+++ .vscode/tasks.json
@@ -1,0 +1,66 @@
+{
+    // See https://go.microsoft.com/fwlink/?LinkId=733558
+    // for the documentation about the tasks.json format
+    "version": "2.0.0",
+    "tasks": [
+        {
+            "label": "Карта зависимостей модуля UI",
+            "type": "shell",
+            "command": "madge",
+            "args": [
+                "UI/Base.ts",
+                "-i",
+                "./ui-libonly.svg",
+                "--ts-config=tsconfig.json",
+                "-x",
+                "\"(^UI/Utils)|(^UI/_utils/)|(^UI/DevtoolsHook)\""
+            ],
+            "isBackground": true,
+            "problemMatcher": []
+        },
+        {
+            "label": "Список циклических зависимостей модуля UI",
+            "type": "shell",
+            "command": "madge",
+            "args": [
+                "UI/Base.ts",
+                "--circular",
+                "--ts-config=tsconfig.json",
+                "-x",
+                "\"(^UI/DevtoolsHook)\""
+            ],
+            "isBackground": true,
+            "problemMatcher": []
+        },
+        {
+            "label": "Карта зависимостей модуля UI/Base",
+            "type": "shell",
+            "command": "madge",
+            "args": [
+                "UI/Base.ts",
+                "-i",
+                "./uibasets-libonly.svg",
+                "--ts-config=tsconfig.json",
+                "-x",
+                "\"(^(.+/_(?!base/).+)|(^UI/DevtoolsHook))\""
+            ],
+            "isBackground": true,
+            "problemMatcher": []
+        },
+        {
+            "label": "Карта зависимостей модуля UI/Vdom",
+            "type": "shell",
+            "command": "madge",
+            "args": [
+                "UI/Vdom.ts",
+                "-i",
+                "./uivdomts-libonly.svg",
+                "--ts-config=tsconfig.json",
+                "-x",
+                "\"(^(.+/_(?!vdom/).+)|(^UI/DevtoolsHook))\""
+            ],
+            "isBackground": true,
+            "problemMatcher": []
+        }
+    ]
+}
