+++ UI/Focus.ts
@@ -10,7 +10,6 @@ import { restoreFocus } from './_focus/RestoreFocus';
 
 import { goUpByControlTree } from 'UI/NodeCollector';
 import * as DefaultOpenerFinder from './_focus/DefaultOpenerFinder';
-import { IDOMEnvironment } from './_focus/Events';
 import * as FocusAttrs from './_focus/FocusAttrs';
 import { IControl } from './_focus/IControl';
 
@@ -20,7 +19,6 @@ export {
    BoundaryElements,
    focus,
    _initFocus,
-   IDOMEnvironment as _IDOMEnvironment,
    IControl as _IControl,
    FocusAttrs as _FocusAttrs,
    nativeFocus,
