+++ UI/_vdom/Synchronizer/interfaces.ts
@@ -1,6 +1,5 @@
 import { Control, IControlOptions } from 'UI/Base';
 import { VNode } from 'Inferno/third-party/index';
-import { IDOMEnvironment } from './resources/DOMEnvironment';
 
 export type TComponentAttrs = Record<string, unknown>;
 
@@ -65,6 +64,10 @@ export interface IControlNode extends IRebuildNode {
 
     // TODO это нужно вынести в расширенные интерфейсы
     _moduleName?: string;  // это добавляет какой то Executor
+
+    // мы не должны зависить от UI/Focus._IDOMEnvironment, поэтому определю тут
+    _rootDOMNode: TModifyHTMLNode;
+    __captureEventHandler: Function;
 }
 
 export interface ICompatableControl {
@@ -95,4 +98,80 @@ export interface IWasabyHTMLElement extends HTMLElement {
     controlNodes: IControlNode[];
     eventProperties: TEventsObject;
     eventPropertiesCnt: number;
+}
+
+export interface IProperties {
+    attributes: Record<string, string>;
+    hooks: Record<string, any>;
+    events: TEventsObject;
+}
+
+
+export type TModifyHTMLNode = HTMLElement & Record<string, any>;
+
+export type TMarkupNodeDecoratorFn = (
+    tagName: string,
+    properties: IProperties,
+    children: any,
+    key: TControlId,
+    controlNode: any,
+    ref: any
+) => VNode[];
+
+export interface IHandlerInfo {
+    handler: (evt: Event) => void;
+    bodyEvent: boolean;
+    processingHandler: boolean;
+    count: number;
+}
+
+export interface IDOMEnvironment {
+    addTabListener(e?: any): void;
+    removeTabListener(e: any): void;
+    destroy(): void;
+
+    _handleFocusEvent(e: any): void;
+    _handleBlurEvent(e: any): void;
+    _handleMouseDown(e: any): void;
+    _handleClick(event: any): void;
+    _handleTouchstart(event: any): void;
+    _handleTouchmove(event: any): void;
+    _handleTouchend(event: any): void;
+    _shouldUseClickByTap(): boolean;
+
+    applyNewVNode(newVNnode: any, rebuildChanges: any, newRootCntNode: any): void;
+    decorateFullMarkup(vnode: VNode, controlNode: any): VNode;
+    getMarkupNodeDecorator(): TMarkupNodeDecoratorFn;
+    getDOMNode(): HTMLElement;
+
+    startEvent(controlNode: any, args: any): any;
+    getHandlerInfo(eventName: string, processingHandler: boolean, bodyEvent: boolean): IHandlerInfo | null;
+    addHandler(eventName: string, isBodyElement: boolean, handler: any, processingHandler: boolean): void;
+    addNativeListener(
+        element: HTMLElement,
+        handler: any,
+        eventName: string,
+        config: any,
+        options?: boolean | AddEventListenerOptions
+    ): void;
+    removeHandler(eventName: string, isBodyElement: boolean, processingHandler: boolean): any;
+    removeNativeListener(element: HTMLElement, handler: EventListener, eventName: string, capture: boolean): any;
+    addCaptureEventHandler(eventName: string, element: IWasabyHTMLElement): any;
+    addCaptureProcessingHandler(eventName: string, method: (event: Event) => void): any;
+    removeCaptureEventHandler(eventName: string, element: IWasabyHTMLElement): void;
+    removeAllCaptureHandlers(): void;
+    removeProcessiingEventHandler(eventName: string): void;
+    _canDestroy(destroyedControl: Control): any;
+
+    queue?: string[];
+
+    _currentDirties: Record<string, number>;
+    _nextDirties: Record<string, number>;
+    activateSubQueue: undefined;
+
+    // FIXME это не должно быть публичным. Найти все ссылки и разобраться
+    _rootDOMNode: TModifyHTMLNode;
+    __captureEventHandler: Function;
+    _rebuildRequestStarted?: boolean;
+    _haveRebuildRequest?: boolean;
 }
