+++ UI/_vdom/Synchronizer/resources/DOMEnvironment.ts
@@ -1,15 +1,16 @@
 /// <amd-module name="UI/_vdom/Synchronizer/resources/DOMEnvironment" />
-// tslint:disable:variable-name
+// tslint:disable:variable-name no-any
 
 import { ArrayUtils } from 'UI/Utils';
-import { _IDOMEnvironment } from 'UI/Focus';
 import { needToBeCompatible } from 'UI/Utils';
 
 import { constants, detection } from 'Env/Env';
 import { Logger, isNewEnvironment } from 'UI/Utils';
-import { Control } from 'UI/Base';
 import { ElementFinder, Events, BoundaryElements, focus, preventFocus, hasNoFocus, goUpByControlTree } from 'UI/Focus';
-import { TControlId, TControlStateCollback, IControlNode, IWasabyHTMLElement, TEventsObject } from '../interfaces';
+import {
+   IDOMEnvironment, TControlStateCollback, IControlNode,
+   IWasabyHTMLElement, TMarkupNodeDecoratorFn, IHandlerInfo, TModifyHTMLNode
+} from '../interfaces';
 
 import { delay } from 'Types/function';
 import { mapVNode } from './VdomMarkup';
@@ -155,8 +156,6 @@ class QueueMixin extends Environment {
       this.queue = null;
    }
 }
-
-type TModifyHTMLNode = HTMLElement & Record<string, any>;
 
 interface IDires {
    [key: string]: number;
@@ -1284,77 +1283,4 @@ if (detection.isIE12 && typeof window !== 'undefined' && typeof document !== 'un
          }
       }, true);
    }
-}
-
-export interface IProperties {
-   attributes: Record<string, string>;
-   hooks: Record<string, any>;
-   events: TEventsObject;
-}
-
-interface IHandlerInfo {
-   handler: (evt: Event) => void;
-   bodyEvent: boolean;
-   processingHandler: boolean;
-   count: number;
-}
-
-type TMarkupNodeDecoratorFn = (
-   tagName: string,
-   properties: IProperties,
-   children: any,
-   key: TControlId,
-   controlNode: any,
-   ref: any
-) => VNode[];
-
-export interface IDOMEnvironment extends _IDOMEnvironment {
-   addTabListener(e?: any): void;
-   removeTabListener(e: any): void;
-   destroy(): void;
-
-   _handleFocusEvent(e: any): void;
-   _handleBlurEvent(e: any): void;
-   _handleMouseDown(e: any): void;
-   _handleClick(event: any): void;
-   _handleTouchstart(event: any): void;
-   _handleTouchmove(event: any): void;
-   _handleTouchend(event: any): void;
-   _shouldUseClickByTap(): boolean;
-
-   applyNewVNode(newVNnode: any, rebuildChanges: any, newRootCntNode: any): void;
-   decorateFullMarkup(vnode: VNode, controlNode: any): VNode;
-   getMarkupNodeDecorator(): TMarkupNodeDecoratorFn;
-   getDOMNode(): HTMLElement;
-
-   startEvent(controlNode: any, args: any): any;
-   getHandlerInfo(eventName: string, processingHandler: boolean, bodyEvent: boolean): IHandlerInfo | null;
-   addHandler(eventName: string, isBodyElement: boolean, handler: any, processingHandler: boolean): void;
-   addNativeListener(
-      element: HTMLElement,
-      handler: any,
-      eventName: string,
-      config: any,
-      options?: boolean | AddEventListenerOptions
-   ): void;
-   removeHandler(eventName: string, isBodyElement: boolean, processingHandler: boolean): any;
-   removeNativeListener(element: HTMLElement, handler: EventListener, eventName: string, capture: boolean): any;
-   addCaptureEventHandler(eventName: string, element: IWasabyHTMLElement): any;
-   addCaptureProcessingHandler(eventName: string, method: (event: Event) => void): any;
-   removeCaptureEventHandler(eventName: string, element: IWasabyHTMLElement): void;
-   removeAllCaptureHandlers(): void;
-   removeProcessiingEventHandler(eventName: string): void;
-   _canDestroy(destroyedControl: Control): any;
-
-   queue?: string[];
-
-   _currentDirties: Record<string, number>;
-   _nextDirties: Record<string, number>;
-   activateSubQueue: undefined;
-
-   // FIXME это не должно быть публичным. Найти все ссылки и разобраться
-   _rootDOMNode: TModifyHTMLNode;
-   __captureEventHandler: Function;
-   _rebuildRequestStarted?: boolean;
-   _haveRebuildRequest?: boolean;
 }
