+++ package.json
@@ -41,5 +41,28 @@
       "requirejs": "2.1.18",
       "serve-static": "1.11.x",
       "wasaby-cli": "git+https://platform-git.sbis.ru/saby/wasaby-cli.git#rc-20.6000"
+   },
+   "madge": {
+      "baseDir": ".",
+      "excludeRegExp": [
+         "^application",
+         "^Compiler",
+         "^Markup",
+         "^UIDemo",
+         "^UITest",
+         "^WmlBuilder",
+         "^UI/DevtoolsHook"
+      ],
+      "fontSize": "26px",
+      "graphVizOptions": {
+         "G": {
+            "rankdir": "LR"
+         }
+      },
+      "detectiveOptions": {
+         "ts": {
+            "skipTypeImports": true
+         }
+      }
    }
 }
